+++ packages/backend/src/event/services/recur/util/recur.util.month.test.ts
@@ -1,5 +1,5 @@
 import { Schema_Event_Core } from "@core/types/event.types";
-import { RRULE } from "../../../../../core/src/constants/core.constants";
+import { RRULE } from "../../../../../../core/src/constants/core.constants";
 import { assembleInstances } from "./recur.util";
 
 describe("Event Recurrence: Month", () => {
