+++ packages/backend/src/event/services/recur/util/recur.util.week.test.ts
@@ -1,6 +1,6 @@
 import dayjs from "dayjs";
 import { Schema_Event_Core } from "@core/types/event.types";
-import { RRULE } from "../../../../../core/src/constants/core.constants";
+import { RRULE } from "../../../../../../core/src/constants/core.constants";
 import { areDatesUnique, haveSharedValues } from "./recur.test.util";
 import { assembleInstances } from "./recur.util";
